tau_env/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.1
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
eval_type_backport==0.2.2
executing==0.8.3
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
invoke==2.2.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
litellm==1.76.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.6
mistralai==1.9.8
multidict==6.6.4
narwhals==2.3.0
nest-asyncio==1.6.0
numpy==2.3.2
openai==1.101.0
packaging==25.0
pandas==2.3.2
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
pip==25.1
platformdirs==4.3.7
prompt-toolkit==3.0.43
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rpds-py==0.27.0
rsa==4.9.1
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.2.0
streamlit==1.49.1
tau_bench==0.1.0
tenacity==9.1.2
termcolor==3.1.0
tiktoken==0.11.0
tokenizers==0.21.4
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.45.1
yarl==1.20.1
zipp==3.23.0
